Refactor for reusing generic component
