+++ builder/generic.go
@@ -1,12 +1,11 @@
-package postaggregation
+package builder
 
 import (
 	"encoding/json"
-	"github.com/grafadruid/go-druid/builder"
 )
 
 type Generic struct {
-	Base
+	Typ    ComponentType
 	Fields map[string]interface{}
 }
 
@@ -16,12 +15,28 @@ func NewGeneric(typ string) *Generic {
 	return g
 }
 
+// Methods for compatibility
+
+func (g *Generic) Type() ComponentType {
+	return g.Typ
+}
+
+func (g *Generic) SetType(typ string) *Generic {
+	g.Typ = typ
+	return g
+}
+
+func (g *Generic) SetName(name string) *Generic {
+	g.SetField("name", name)
+	return g
+}
+
+// Generic methods
+
 func (g *Generic) SetField(name string, value interface{}) *Generic {
 	switch name {
 	case "type":
-		g.SetType(value.(builder.ComponentType))
-	case "name":
-		g.SetName(value.(string))
+		g.SetType(value.(ComponentType))
 	default:
 		g.Fields[name] = value
 	}
@@ -29,6 +44,11 @@ func (g *Generic) SetField(name string, value interface{}) *Generic {
 }
 
 func (g *Generic) SetFields(fields map[string]interface{}) *Generic {
+	if typ, present := fields["type"]; present {
+		g.SetType(typ.(ComponentType))
+		delete(fields, "type")
+	}
+
 	g.Fields = fields
 	return g
 }
@@ -45,14 +65,10 @@ func (g *Generic) MarshalJSON() ([]byte, error) {
 	if g.Typ != "" {
 		g.Fields["type"] = g.Typ
 	}
-	if g.Name != "" {
-		g.Fields["name"] = g.Name
-	}
 
 	data, err := json.Marshal(&g.Fields)
 
 	delete(g.Fields, "type")
-	delete(g.Fields, "name")
 
 	return data, err
 }
@@ -61,14 +77,10 @@ func (g *Generic) UnmarshalJSON(bytes []byte) error {
 	err := json.Unmarshal(bytes, &g.Fields)
 
 	if typ, present := g.Fields["type"]; present {
-		g.Typ = typ.(builder.ComponentType)
-	}
-	if name, present := g.Fields["name"]; present {
-		g.Name = name.(string)
+		g.Typ = typ.(ComponentType)
 	}
 
 	delete(g.Fields, "type")
-	delete(g.Fields, "name")
 
 	return err
 }
